Move DirIndicator class
